The source adaptation to compile against the current mintlib and cflib CVS versions.
	- FreeMiNT compatible Makefile structure (stolen)
	- int -> short data type conversion
	- set_int() -> set_short() cflib usage
	- LINEENDING enum value names conversion (crosscompilation conflicts)
	- EOS defined to '\0' in types.h
	- #define CDECL cdecl if PureC else an empty define
	- several type scale adjustments (e.g. switch(short) { case 0xFEFD: }
		to make gcc not to warn switch( unsigned short )

Compiles, runs but can't load a text file :(

+++ src/aktion.h
@@ -1,7 +1,7 @@
 #ifndef _qed_aktion_h_
 #define _qed_aktion_h_
 
-extern int		akt_handle;
+extern short		akt_handle;
 
 extern void		start_aktion		(char *str, bool inter, long max);
 extern bool		do_aktion			(char *str, long value);
