+++ src/ausgabe.h
@@ -2,23 +2,23 @@
 #define _qed_ausgabe_h_
 
 /* Liefert die interne Position */
-extern int		inter_pos		(int x, ZEILEP a, bool tab, int tab_size);
+extern short		inter_pos		(short x, ZEILEP a, bool tab, short tab_size);
 
 /* Liefert die L„nge auf dem Bild */
-extern int		bild_len			(ZEILEP a, bool tab, int tab_size);
+extern short		bild_len			(ZEILEP a, bool tab, short tab_size);
 
 /* Liefert die Position auf dem Bild */
-extern int		bild_pos			(int x, ZEILEP a, bool tab, int tab_size);
+extern short		bild_pos			(short x, ZEILEP a, bool tab, short tab_size);
 
-extern void		fill_area		(int x, int y, int w, int h, int color);
-extern int		out_s				(int x, int y, int w, char *str);
-extern int		out_sb			(int x, int y, int w, char *str);
+extern void		fill_area		(short x, short y, short w, short h, short color);
+extern short		out_s				(short x, short y, short w, char *str);
+extern short		out_sb			(short x, short y, short w, char *str);
 
 extern void		head_out			(WINDOWP window, TEXTP t_ptr);
-extern void		line_out			(WINDOWP window, TEXTP t_ptr, int wy);
+extern void		line_out			(WINDOWP window, TEXTP t_ptr, short wy);
 extern void		bild_out			(WINDOWP window, TEXTP t_ptr);
 extern void		bild_blkout		(WINDOWP window, TEXTP t_ptr, long z1, long z2);
-extern int		cursor_xpos		(TEXTP t_ptr, int pos);
+extern short		cursor_xpos		(TEXTP t_ptr, short pos);
 extern void		cursor			(WINDOWP window, TEXTP t_ptr);
 
 extern void 	set_drawmode	(void);
