+++ src/edit.h
@@ -1,7 +1,7 @@
 #ifndef _qed_edit_h_
 #define _qed_edit_h_
 
-extern int		edit_type;
+extern short		edit_type;
 
 
 extern void		absatz_edit		(void);
@@ -10,17 +10,17 @@ extern void		pos_korr			(WINDOWP window, TEXTP t_ptr);
 extern void		blink_edit		(void);
 extern void		offblink_edit	(void);
 extern void		onblink_edit	(void);
-extern int		new_edit			(void);
-extern int		load_edit		(char *filename, bool bin);
-extern void		close_edit		(char *filename, int flag);
+extern short		new_edit			(void);
+extern short		load_edit		(char *filename, bool bin);
+extern void		close_edit		(char *filename, short flag);
 
-extern bool		info_edit		(int icon);
+extern bool		info_edit		(short icon);
 
-extern void		make_chg			(int link, int change, long ypos);
+extern void		make_chg			(short link, short change, long ypos);
 extern void		restore_edit	(void);
 
-extern void		cursor_on		(int wHandle);
-extern void		cursor_off		(int wHandle);
+extern void		cursor_on		(short wHandle);
+extern void		cursor_off		(short wHandle);
 
 /* Infozeile an- bzw. abschalten */
 extern void		set_info			(TEXTP t_ptr, char *str);
