+++ src/find.h
@@ -11,7 +11,7 @@
 
 extern bool 	s_grkl, s_quant, s_wort, s_vorw, s_global, s_round,
 					ff_rekursiv;
-extern int		r_modus, rp_box_x, rp_box_y;
+extern short		r_modus, rp_box_x, rp_box_y;
 extern char		r_str[], s_str[], 
 					s_history[HIST_ANZ][HIST_LEN+1],
 					r_history[HIST_ANZ][HIST_LEN+1],
@@ -21,12 +21,12 @@ extern UMLAUTENCODING	umlaut_from,
 								umlaut_to;
 
 
-extern int		start_find		(TEXTP t_ptr, bool quiet);
-extern int		start_replace	(TEXTP t_ptr);
-extern int		do_next			(TEXTP t_ptr);
+extern short		start_find		(TEXTP t_ptr, bool quiet);
+extern short		start_replace	(TEXTP t_ptr);
+extern short		do_next			(TEXTP t_ptr);
 extern void		find_selection	(TEXTP t_ptr);
 
-extern bool		filematch		(char *filename, char *m, int fs_typ);
+extern bool		filematch		(char *filename, char *m, short fs_typ);
 /*
  * Testet auf Gleichheit von Dateiname und Muster.
  * Beachtet dabei fs_case_sens().
@@ -38,7 +38,7 @@ extern void		change_umlaute	(TEXTP t_ptr);
 
 /* Dialoge */
 
-extern int		replace_dial	(void);
+extern short		replace_dial	(void);
 extern bool		findfile_dial	(char *path, bool in_prj);
 extern bool		umlaut_dial		(void);
